fix slow selection, add del verification (include # points)

+++ src/Gui/mnuRibbon.py
@@ -409,8 +409,13 @@ class mnuRibbon(RB.RibbonBar):
         event.Skip()
 
     def onEditDelPoint(self, event):
-        self.parent.getRecordService().delete_points()
-        Publisher.sendMessage(("updateValues"), event=event)
+        numPoints=len(self.parent.getRecordService().get_filtered_points())
+        val = wx.MessageBox("You have chosen to delete the %s selected points.\nDo you want to continue?"% numPoints,
+                               'Deleting Points',
+                               wx.YES_NO|wx.ICON_QUESTION)
+        if val == 2:#wx.ID_YES:
+            self.parent.getRecordService().delete_points()
+            Publisher.sendMessage(("updateValues"), event=event)
         event.Skip()
 
     def onRestore(self, event):
